feat: Add createUser function to patient.actions
This commit adds the `createUser` function to the `patient.actions` module. The function is responsible for creating a new user in the backend using the provided user data. It utilizes the `users.create` method from the `node-appwrite` library to create the user and returns the newly created user object. This addition enhances the functionality of the application by enabling the creation of new users and their registration process.